optimisation de moneyPal(invitation et transaction OrangeMoney Terminees)

+++ MoneyPal/app/src/main/java/com/example/workstation/moneypal/entities/DetailOperation.kt
@@ -6,7 +6,7 @@ import java.util.*
 data class DetailOperation(
     val operation: Operation?,
     val date:Date?,
-    var amount:Int,
+    var amount:Double,
     val smsObject: SmsObject?) :Serializable {
-    constructor() : this(null,null,0,null)
+    constructor() : this(null,null,0.0,null)
 }
